[1569] Only delegate semantic deletion to the element's Delete Tool
The behavior of "Delete from diagram", when available (on
unsynchronized elements) is hard-coded to always perform the expected
behavior, nothing more and nothing less.

Delete Tool (default or user-supplied) only need to consider the case
of semantic deletion.

Bug: https://github.com/eclipse-sirius/sirius-components/issues/1569
